Finished warehouse_setup 1.0
Not the prettiest but still works pretty well all things considered. It can now read the base file, create and save dimension tables from what you give it, and makes the fact table at the end. It certainly still needs work, but it works for now so I can move on to experimenting with neural networks.